third_party/blink/web_tests/external/wpt/editing/run/justifycenter_6001-last-expected.txt

This is a testharness.js-based test.
Found 20 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<div style=text-align:center><p>foo</div> <p>[bar]</p> <div style=text-align:center><p>baz</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foo</p> <p>bar</p> <p>baz</p></div>" but got "<div style=\\"text-align:center\\"><p>foo</p></div> <p style=\\"text-align:center\\">bar</p> <div style=\\"text-align:center\\"><p>baz</p></div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<div style=text-align:center><p>foo</div> <p>[bar]</p> <div style=text-align:center><p>baz</div>" queryCommandState("stylewithcss") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<div style=text-align:center><p>foo</div> <p>[bar]</p> <div style=text-align:center><p>baz</div>" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["defaultparagraphseparator","div"],["justifycenter",""]] "<div align=center><p>foo</div> <p>[bar]</p> <div align=center><p>baz</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div align=\\"center\\"><p>foo</p> <p>bar</p> <p>baz</p></div>" but got "<div align=\\"center\\"><p>foo</p></div> <p style=\\"text-align:center\\">bar</p> <div align=\\"center\\"><p>baz</p></div>"
[FAIL] [["defaultparagraphseparator","p"],["justifycenter",""]] "<div align=center><p>foo</div> <p>[bar]</p> <div align=center><p>baz</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div align=\\"center\\"><p>foo</p> <p>bar</p> <p>baz</p></div>" but got "<div align=\\"center\\"><p>foo</p></div> <p style=\\"text-align:center\\">bar</p> <div align=\\"center\\"><p>baz</p></div>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<center><p>foo</p></center><div style=\\"text-align:center\\"> <p>bar</p> </div><center><p>baz</p></center>" but got "<center><p>foo</p></center> <p style=\\"text-align:center\\">bar</p> <center><p>baz</p></center>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<center><p>foo</p></center><div style=\\"text-align:center\\"> <p>bar</p> </div><center><p>baz</p></center>" but got "<center><p>foo</p></center> <p style=\\"text-align:center\\">bar</p> <center><p>baz</p></center>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<center><p>foo</p></center><div style=\\"text-align:center\\"> <p>bar</p> </div><center><p>baz</p></center>" but got "<center><p>foo</p></center> <p style=\\"text-align:center\\">bar</p> <center><p>baz</p></center>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<center><p>foo</p></center><div style=\\"text-align:center\\"> <p>bar</p> </div><center><p>baz</p></center>" but got "<center><p>foo</p></center> <p style=\\"text-align:center\\">bar</p> <center><p>baz</p></center>"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "center" but got "-webkit-center"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "center" but got "-webkit-center"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "center" but got "-webkit-center"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "-webkit-center"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "center" but got "-webkit-center"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "center" but got "-webkit-center"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "center" but got "-webkit-center"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "center" but got "-webkit-center"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "{ <span contenteditable=\\"false\\">A</span> ; <span contenteditable=\\"false\\">B</span> ; <span contenteditable=\\"false\\">C</span> }" checks for modifications to non-editable content
  assert_equals: Wrapper div must have only one attribute (<div contenteditable="">), but has more (<div contenteditable="" style="text-align: center;">) expected 1 but got 2
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "{ <span contenteditable=\\"false\\">A</span> ; <span contenteditable=\\"false\\">B</span> ; <span contenteditable=\\"false\\">C</span> }" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div style=\\"text-align:center\\"></div> <span contenteditable=\\"false\\">A</span> ; <span contenteditable=\\"false\\">B</span> ; <span contenteditable=\\"false\\">C</span> " not in array ["<div style=\\"text-align:center\\"> <span contenteditable=\\"false\\">A</span> ; <span contenteditable=\\"false\\">B</span> ; <span contenteditable=\\"false\\">C</span> </div>", " <div style=\\"text-align:center\\"><span contenteditable=\\"false\\">A</span> ; <span contenteditable=\\"false\\">B</span> ; <span contenteditable=\\"false\\">C</span></div> ", "<div style=\\"text-align:center\\"><span contenteditable=\\"false\\">A</span> ; <span contenteditable=\\"false\\">B</span> ; <span contenteditable=\\"false\\">C</span></div>"]
Harness: the test ran to completion.
